+++ src/NCronJob/Scheduler/JobQueueManager.cs
@@ -47,7 +47,15 @@ internal sealed class JobQueueManager : IDisposable
         {
             if (jobQueues.TryRemove(queueName, out var jobQueue))
             {
-                // TODO: Should we first cancel the potentially existing JobRuns here?
+                foreach (var job in jobQueue)
+                {
+                    if (!job.IsCancellable)
+                    {
+                        continue;
+                    }
+
+                    job.NotifyStateChange(JobStateType.Cancelled);
+                }
 
                 jobQueue.Clear();
                 jobQueue.CollectionChanged -= CallCollectionChanged;